net: is_skb_forwardable: check the size of GSO segments
is_skb_forwardable attempts to detect if a packet is too large to
be sent to the destination device. However, this test does not
consider GSO skbs, and it is possible that a GSO skb, when
segmented, will be larger than the device can transmit.

Use skb_gso_validate_mac_len to check.

Signed-off-by: Daniel Axtens <dja@axtens.net>